Version 1.2 of the phmo export script.
The new version has improved performance and is easier to use. The user
no longer has to apply transformations manually to objects in the scene,
this is now done automatically without affecting the objects in the
scene. It also no longer matters which way the normals are facing on the
models to be exported. The script will determine the outward facing
direction for planes.